Remove "new utilities available" from comparisons
I do not have "Managing Remote Connections" in my UNSTABLE install
I may need to install/configure something to enable it. If so,
this may be better mentioned in a different section with further
details, and/or point to the sysadm manual?

Moved Task Manager to the "Converted to SysADM" section in the
rewrite for that part of the TruOS comparison to FreeBSD/PC-BSD.